# mapper-service/openapi.yaml
openapi: 3.0.3
info:
  title: Mapper Service API
  description: |
    The Mapper Service provides core mapping functionality, model serving, and 
    response generation within the llama-mapper microservice architecture. It 
    handles the transformation of analysis results into canonical taxonomy and 
    compliance framework mappings.
  version: 1.0.0
  contact:
    name: Llama Mapper Team
    email: team@llamapper.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8002
    description: Development server
  - url: https://api.llamapper.com/mapper
    description: Production server
paths:
...
  /api/v1/training/jobs:
    post:
      summary: Submit training job
      operationId: submitTrainingJob
      tags:
        - Training
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainingJobRequest'
      responses:
        '202':
          description: Training job submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingJobResponse'

    get:
      summary: List training jobs
      operationId: listTrainingJobs
      tags:
        - Training
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
      responses:
        '200':
          description: List of training jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainingJob'
  /api/v1/training/jobs/{jobId}:
    get:
      summary: Get training job status
      operationId: getTrainingJobStatus
      tags:
        - Training
      security:
        - ApiKeyAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Training job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingJob'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    TenantId:
      name: X-Tenant-ID
      in: header
      required: true
      schema:
        type: string
      description: Tenant identifier for multi-tenancy
...
  schemas:
...
    TrainingJobRequest:
      type: object
      required:
        - job_name
        - base_model
        - training_type
        - training_data_path
        - configuration
      properties:
        job_name:
          type: string
        base_model:
          type: string
        training_type:
          type: string
          enum: [lora, full_finetune, qlora]
        training_data_path:
          type: string
        validation_data_path:
          type: string
        configuration:
          type: object
    TrainingJobResponse:
      type: object
      required:
        - job_id
        - status
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        estimated_completion_time:
          type: string
          format: date-time
    TrainingJob:
      type: object
      required:
        - job_name
        - base_model
        - training_type
        - status
      properties:
        id:
          type: string
          format: uuid
        job_name:
          type: string
        base_model:
          type: string
        training_type:
          type: string
          enum: [lora, full_finetune, qlora]
        training_data_path:
          type: string
        validation_data_path:
          type: string
        output_model_path:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        configuration:
          type: object
        metrics:
          type: object
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        training_time_minutes:
          type: integer
        error_message:
          type: string
...